+++ SubCodes/UnKronPolicyIndexes/UnKronPolicyIndexes4_z_e.m
@@ -16,8 +16,8 @@ l_daprime4=length(n_daprime4);
 
 divisors1=cumprod([1,n_daprime1(1:end-1)])';   % [l_daprime1,1]
 divisors2=cumprod([1,n_daprime2(1:end-1)])';   % [l_daprime2,1]
-divisors3=cumprod([1,double(n_daprime3(1:end-1))])';   % [l_daprime3,1]
-divisors4=cumprod([1,double(n_daprime4(1:end-1))])';   % [l_daprime4,1]
+divisors3=cumprod([1,n_daprime3(1:end-1)])';   % [l_daprime3,1]
+divisors4=cumprod([1,n_daprime4(1:end-1)])';   % [l_daprime4,1]
 
 if vfoptions.gridinterplayer==1
     Policy=[mod(floor(double(PolicyKron(1,:)-1)./divisors1),n_daprime1(:))+1;
